Map Game
I call it MAP Game because there is a map you will see at the beginning of the game. On the left bottom of the map, you will see your status of your "hero". Try to press "Space" to get into each room of the map. Some of them are with enemies and some of them have rewards like money etc. Try to finish the whole map!!